+++ app/client/src/components/editorComponents/ActionCreator/constants.ts
@@ -1,0 +1,18 @@
+export const APPSMITH_GLOBAL_FUNCTIONS = {
+  navigateTo: "navigateTo",
+  showAlert: "showAlert",
+  showModal: "showModal",
+  closeModal: "closeModal",
+  storeValue: "storeValue",
+  download: "download",
+  copyToClipboard: "copyToClipboard",
+  resetWidget: "resetWidget",
+  setInterval: "setInterval",
+  clearInterval: "clearInterval",
+};
+
+export const APPSMITH_NAMESPACED_FUNCTIONS = {
+  getGeolocation: "appsmith.geolocation.getCurrentPosition",
+  watchGeolocation: "appsmith.geolocation.watchPosition",
+  stopWatchGeolocation: "appsmith.geolocation.clearWatch",
+};
