--- app/client/src/workers/DependencyMap/constants.ts
@@ -1,20 +1,0 @@
-import { APPSMITH_GLOBAL_FUNCTIONS } from "components/editorComponents/ActionCreator/Fields";
-import {
-  DEDICATED_WORKER_GLOBAL_SCOPE_IDENTIFIERS,
-  JAVASCRIPT_KEYWORDS,
-} from "constants/WidgetValidation";
-import { extraLibrariesNames } from "utils/DynamicBindingUtils";
-
-/**
- * Identifiers which can not be valid names of entities and are not dynamic in nature.
- * therefore should be removed from the list of references extracted from code.
- * NB: DATA_TREE_KEYWORDS in app/client/src/constants/WidgetValidation.ts isn't included, although they are not valid entity names,
- * they can refer to potentially dynamic entities.
- * Eg. "appsmith"
- */
-export const invalidEntityIdentifiers: Record<string, unknown> = {
-  ...JAVASCRIPT_KEYWORDS,
-  ...APPSMITH_GLOBAL_FUNCTIONS,
-  ...DEDICATED_WORKER_GLOBAL_SCOPE_IDENTIFIERS,
-  ...extraLibrariesNames,
-};
