+++ app/client/src/workers/DependencyMap/utils.ts
@@ -5,11 +5,16 @@ import {
   EvalError,
   DependencyMap,
   getDynamicBindings,
+  extraLibrariesNames,
 } from "utils/DynamicBindingUtils";
 import { extractInfoFromCode } from "@shared/ast";
 import { convertPathToString, isWidget } from "../evaluationUtils";
-import { invalidEntityIdentifiers } from "./constants";
 import { DataTreeWidget } from "entities/DataTree/dataTreeFactory";
+import {
+  DEDICATED_WORKER_GLOBAL_SCOPE_IDENTIFIERS,
+  JAVASCRIPT_KEYWORDS,
+} from "constants/WidgetValidation";
+import { APPSMITH_GLOBAL_FUNCTIONS } from "components/editorComponents/ActionCreator/constants";
 
 /** This function extracts validReferences and invalidReferences from a binding {{}}
  * @param script
@@ -154,4 +159,18 @@ export function listTriggerFieldDependencies(
 export const mergeArrays = <T>(currentArr: T[], updateArr: T[]): T[] => {
   if (!currentArr) return updateArr;
   return union(currentArr, updateArr);
+};
+
+/**
+ * Identifiers which can not be valid names of entities and are not dynamic in nature.
+ * therefore should be removed from the list of references extracted from code.
+ * NB: DATA_TREE_KEYWORDS in app/client/src/constants/WidgetValidation.ts isn't included, although they are not valid entity names,
+ * they can refer to potentially dynamic entities.
+ * Eg. "appsmith"
+ */
+const invalidEntityIdentifiers: Record<string, unknown> = {
+  ...JAVASCRIPT_KEYWORDS,
+  ...APPSMITH_GLOBAL_FUNCTIONS,
+  ...DEDICATED_WORKER_GLOBAL_SCOPE_IDENTIFIERS,
+  ...extraLibrariesNames,
 };
